% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_net2plot.R
\name{create_net2plot}
\alias{create_net2plot}
\title{This function create an igraph network object for visualization puroposes}
\usage{
create_net2plot(
  g_net,
  input_m,
  gf_prop,
  ann_net_b,
  frag_pattern = "F",
  ff_net = NULL,
  ff_prop = NULL
)
}
\arguments{
\item{g_net}{matrix, in each row there are starting and ending nodes of one edge 
in the matrix. This matrix represent a gene-gene or gene-fragment network.}

\item{input_m}{vector of dimension nX1 representing the transcriptional profile befor propagation}

\item{gf_prop}{vector of dimension nX1 representing the transcriptional profile after 
propagation using a gene-gene or gene-fragment network}

\item{ann_net_b}{matrix, for each row presents the gene identifier, the chromosome in which the gene is,
the starting end ending position in the sequence.}

\item{frag_pattern}{string, initial character of the fragments name}

\item{ff_net}{g_net matrix, in each row there are starting and ending nodes of one edge 
in the matrix. This matrix represent a fragment-fragment network.}

\item{ff_prop}{vector of dimension nX1 representing the transcriptional profile after
propagation using a gene-fragment network}
}
\value{
igraph object
}
\description{
This function create an igraph network object for visualization puroposes
}
