% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example}
\alias{example}
\title{Example data for propagation visiualization}
\format{
List
}
\usage{
data(example)
}
\description{
Example data for propagation visiualization
}
\examples{
data(example)
}
\keyword{datasets}
