% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_GUI.R
\name{start_GUI}
\alias{start_GUI}
\title{This function starts a shiny gui for the visualization of propagation results}
\usage{
start_GUI(net1, ann_net_b, chr_len, example = FALSE)
}
\arguments{
\item{net1}{igraph object to visualize}

\item{ann_net_b}{dataframe, specifies the genes positions (chromosome, start 
and end position)}

\item{chr_len}{dataframe, specifies the chromosomes (first column) and their length
(second column)}

\item{example}{boolean, show 5 genes to select as an example}
}
\description{
This function starts a shiny gui for the visualization of propagation results
}
